move ConfigWriter classes into a module
